pyproject.toml
[tool.poetry]
name = "hierarchy"
version = "0.1.0"
description = ""
authors = ["Joseph Brown <joseph.d.brown94@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
